pyproject.toml
[project]
name = "research-agent-with-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==25.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "aiosqlite==0.22.1",
    "altair==6.0.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "appnope==0.1.4 ; sys_platform == 'darwin'",
    "asttokens==3.0.1",
    "attrs==25.4.0",
    "beautifulsoup4>=4.14.3",
    "blinker==1.9.0",
    "cachetools==7.0.2",
    "certifi==2026.2.25",
    "cffi==2.0.0 ; implementation_name == 'pypy' or platform_python_implementation != 'PyPy'",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cryptography==46.0.5",
    "custom-azureopenai-langchain",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.20",
    "decorator==5.2.1",
    "distro==1.9.0",
    "executing==2.2.1",
    "fastapi==0.135.1",
    "filetype==1.2.0",
    "fitz>=0.0.1.dev2",
    "fonttools==4.61.1",
    "frozenlist==1.8.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "google-auth==2.48.0",
    "google-genai==1.65.0",
    "greenlet==3.3.2 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "ipykernel==7.2.0",
    "ipython==9.10.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.8.0",
    "jupyter-core==5.9.1",
    "kiwisolver==1.4.9",
    "langchain==1.2.10",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.17",
    "langchain-google-genai==4.2.1",
    "langchain-openai==1.1.10",
    "langchain-text-splitters==1.1.1",
    "langgraph==1.0.10",
    "langgraph-checkpoint==4.0.1",
    "langgraph-checkpoint-sqlite==3.0.3",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.9",
    "langsmith==0.7.11",
    "markitdown>=0.1.5",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "matplotlib==3.10.8",
    "matplotlib-inline==0.2.1",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "narwhals==2.17.0",
    "nest-asyncio==1.6.0",
    "numpy==2.4.2",
    "openai==2.24.0",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "pandas==2.3.3",
    "parso==0.8.6",
    "pdfplumber>=0.11.9",
    "pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'",
    "pillow==12.1.1",
    "pip==26.0.1",
    "platformdirs==4.9.2",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "psutil==7.2.2",
    "ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'",
    "pure-eval==0.2.3",
    "pyarrow==23.0.1",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0 ; (implementation_name != 'PyPy' and platform_python_implementation != 'PyPy') or (implementation_name == 'pypy' and platform_python_implementation == 'PyPy')",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pymupdf>=1.27.1",
    "pyparsing==3.3.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "pytz==2026.1.post1",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "redis==7.2.1",
    "referencing==0.37.0",
    "regex==2026.2.28",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.30.0",
    "rsa==4.9.1",
    "scrapy>=2.14.1",
    "serpapi>=0.1.5",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.48",
    "sqlite-vec==0.1.6",
    "stack-data==0.6.3",
    "starlette==0.52.1",
    "streamlit==1.55.0",
    "tenacity==9.1.4",
    "tiktoken==0.12.0",
    "toml==0.10.2",
    "tornado==6.5.4",
    "tqdm==4.67.3",
    "trafilatura>=2.0.0",
    "traitlets==5.14.3",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "uvicorn==0.41.0",
    "watchdog==6.0.0 ; sys_platform != 'darwin'",
    "wcwidth==0.6.0",
    "websockets==16.0",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zstandard==0.25.0",
]

[tool.uv.sources]
custom-azureopenai-langchain = { path = "../../../../Projects/python-package/custom_azureopenai_langchain-0.1.0-py3-none-any.whl" }
